optimize submodel
