+++ template.pybind.cpp
@@ -1,8 +1,8 @@
-#include "pybind_model_registry.h"
+#include "pybind_module_registry.h"
 #include <pybind11/pybind11.h>
 #include "{{ util.module_cfg_header_name(module) }}"
 
-ONEFLOW_PYBIND11_MODEL({{ python_module_name }}, m) {
+ONEFLOW_PYBIND11_MODULE("{{ util.module_get_python_module_path(module) }}", m) {
 {% if util.module_has_package(module) %}
   using namespace {{ "::".join(util.module_package_list(module)) }}::cfg;
 {% else %}
